fix: address all PR #140 review comments
Review fixes:
- #1: Delete e2e_coverage_test.go (test padding), move valuable tests
  to stream_parser_test.go, vecli_stream_test.go, auto_resume_test.go
- #3: Rewrite auth.spec.ts with localhost bypass disclaimer, simplify
  to 3 tests verifying authenticated state only
- #5: Remove non-existent .quick-send-hint selector from chat.page.ts
- #6: Fix coverage fixture to check projectName instead of browserName
- #7: Change MockAIBackend to use select/time.After pattern for instant
  cancellation instead of select/default + time.Sleep
- #9: Trim design doc from 1289 to 199 lines, focusing on architecture
  decisions and key design, removing code that duplicates source files

Reviews #2, #4, #8 were already resolved in current codebase.
